separate platform GPIO implementations
Move Arduino and Linux/Demo GPIO implementations into the platform boundary while retaining gpio.h as a compatibility include. Wire platform sources into Makefile and PlatformIO builds without changing callers or hardware behavior.
